@@ -14,7 +14,6 @@ classifier =
         License :: OSI Approved :: MIT License
         Operating System :: POSIX :: Linux
         Programming Language :: Python :: 3
-        Programming Language :: Python :: 3.8
         Programming Language :: Python :: 3.9
         Programming Language :: Python :: 3.10
         Programming Language :: Python :: 3.11
@@ -33,7 +32,7 @@ install_requires =
 packages = find:
 package_dir =
     =src
-python_requires = >=3.8
+python_requires = >=3.9
 zip_safe = False
 
 [options.package_data]
