@@ -8,11 +8,11 @@ import nox
 
 nox.options.sessions = ["lint", "test-dist"]
 
-PYTHON_ALL_VERSIONS = ["3.8", "3.9", "3.10", "3.11", "3.12", "3.13"]
+PYTHON_ALL_VERSIONS = ["3.9", "3.10", "3.11", "3.12", "3.13"]
 RUNNING_CI = "TRAVIS" in os.environ or "GITHUB_ACTIONS" in os.environ
 
 
-@nox.session(python=["3.8"], reuse_venv=True)
+@nox.session(python=["3.9"], reuse_venv=True)
 def lint(session: nox.Session) -> None:
     """
     Run linters on the codebase.
